Debugged the engine
Note that the error was in transforming bindings back to the Engine. As
it generated a new Binding, it has a new Hash, and thus it could not
fire. There were two possibilities to solve this:
- Option 1: change hash generation of Binding
- Option 2: Store the PNIDBinding - Binding mapping

Option 1:
Pro:
 - easier in creating and matching bindings generated in different
 locations
 - makes bindings 'unique'
con
 - need to be altered in pnidprocessor project
 - code not accessible for this course

Option 2:
pro:
- easy to do with a hashmap
con:
- only works if we go 'one way': first from binding to pnidbinding, and
then back

Decision:

In the context of: binding mapping
Facing           : basic functionality
We decided       : to use a hash map
and not          : implement a new hash function
because          : it is easier to implement, and the other code base is
                   unavailable
accepting        : that it is not the best solution, and only works one
		   way.